// scraper/src/app/config/account/account-country.type.ts
export type AccountCountry =
	| "GB"
	| "NB"
	| "AT"
	| "BY"
	| "BE"
	| "BG"
	| "HR"
	| "CZ"
	| "DK"
	| "EE"
	| "FI"
	| "FR"
	| "DE"
	| "GR"
	| "HU"
	| "IS"
	| "IE"
	| "IT"
	| "LV"
	| "LT"
	| "LU"
	| "MC"
	| "NL"
	| "NO"
	| "PL"
	| "PT"
	| "RO"
	| "RU"
	| "SK"
	| "SI"
	| "ES"
	| "SE"
	| "CH"
	| "TR"
	;